requirements.txt
Django==4.1
psycopg2==2.8.6
djangorestframework==3.13.1
debugpy==1.5.1
drf-yasg==1.21.4